+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 * Removed arena.h from public header files.
 * By default, checksums are verified on every read from database
+* Added is_manual_compaction to CompactionFilter::Context
 
 ## 2.7.0 (01/28/2014)